refactor: Split Cli class
